+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 + Upgrade dependencies
 + Add update DNS
 + Refactor instructions and logger
++ Add handle root zone if subdomain is missing
 
 ## 0.3.0
 + Check if public IP is different from DNS record